+++ src/plugin.js
@@ -9,13 +9,16 @@
  *
  * @property {string} folder
  * A path to the folder (relative to `root`).
+ * Namespace: all.
  *
  * @property {(absolutePath: string) => boolean} method
  * A method that accepts an item path (`root` + folderPath + fileName) and
  * returns value that indicates should this item be removed or not.
+ * Namespace: all.
  *
  * @property {boolean} recursive
  * Apply this method to all items in subdirectories.
+ * Namespace: all.
  */
 
 /**
@@ -26,18 +29,22 @@
  * A root directory.
  * Not absolute paths will be appended to this.
  * Defaults to `.` (where `package.json` and `node_modules` are located).
+ * Namespace: all.
  *
  * @property {string[]} include
  * A folders and files for removing.
  * Defaults to `[]`.
+ * Namespace: all.
  *
  * @property {string[]} exclude
  * A folders and files for excluding.
  * Defaults to `[]`.
+ * Namespace: all.
  *
  * @property {TestObject[]} test
  * A folders for testing.
  * Defaults to `[]`.
+ * Namespace: all.
  *
  * @property {(
  *  absoluteFoldersPaths: string[],
@@ -48,8 +55,10 @@
  * will be passed into this function.
  * If returned value is `true`, then
  * remove process will be canceled.
- * Will be not called if `emulate` is on.
+ * Will be not called if items for removing
+ * not found, `emulate: true` or `skipFirstBuild: true`.
  * Defaults to `undefined`.
+ * Namespace: all.
  *
  * @property {(
  *  absoluteFoldersPaths: string[],
@@ -58,48 +67,69 @@
  * If specified, will be called after removing.
  * Absolute paths of folders and files that have been removed
  * will be passed into this function.
+ * Will be not called if `emulate: true` or `skipFirstBuild: true`.
  * Defaults to `undefined`.
+ * Namespace: all.
  *
  * @property {boolean} trash
  * Move folders and files to trash (recycle bin)
  * instead of permanent deleting.
  * Defaults to `false`.
+ * Namespace: all.
  *
  * @property {boolean} log
  * Print messages of "info" level
  * (example: "Which folders or files have been removed").
  * Defaults to `true`.
+ * Namespace: all.
  *
  * @property {boolean} logWarning
  * Print messages of "warning" level
  * (example: "An items for removing not found").
  * Defaults to `true`.
+ * Namespace: all.
  *
  * @property {boolean} logError
  * Print messages of "error" level
  * (example: "No such file or directory").
  * Defaults to `false`.
+ * Namespace: all.
  *
  * @property {boolean} logDebug
  * Print messages of "debug" level
  * (used for debugging).
  * Defaults to `false`.
+ * Namespace: all.
  *
  * @property {boolean} emulate
  * Emulate remove process.
  * Print which folders and files will be removed without actually removing them.
  * Ignores `log` parameter.
  * Defaults to `false`.
+ * Namespace: all.
  *
  * @property {boolean} allowRootAndOutside
  * Allow removing of the `root` directory or outside the `root` directory.
  * It's kind of safe mode.
  * Don't turn it on if you don't know what you actually do!
  * Defaults to `false`.
+ * Namespace: all.
  *
  * @property {PathType} _pathType
  * Type of `path` module.
  * Defaults to `''` (will be selected based on OS).
+ * Namespace: all.
+ *
+ * @property {boolean} skipFirstBuild
+ * First build will be skipped.
+ * Defaults to `false`.
+ * Namespace: `watch`.
+ *
+ * @property {boolean} beforeForFirstBuild
+ * For first build `before` parameters will be applied,
+ * for subsequent builds `watch` parameters will be applied.
+ * Defaults to `false`.
+ * Namespace: `watch`.
  */
 
 /**
@@ -124,6 +154,19 @@
 /**
  * @typedef {Object} Compilation
  * Webpack environment.
+ */
+
+/**
+ * @typedef {{
+ *  name: {
+ *      v4: string;
+ *      v3: string;
+ *  };
+ *  callsCount: int;
+ *  increaseCallsCount: () => void;
+ *  getParams: () => RemovingParameters;
+ * }} WebpackHook
+ * Information about Webpack hook.
  */
 
 //#endregion
@@ -174,6 +217,7 @@ class Plugin {
 
         /** @type {RemovingParameters} */
         const defaultParams = {
+            /* Common parameters */
             root: this.path.path.resolve('.'),
             include: [],
             exclude: [],
@@ -187,7 +231,11 @@ class Plugin {
             logDebug: false,
             emulate: false,
             allowRootAndOutside: false,
-            _pathType: ''
+            _pathType: '',
+
+            /* `watch` parameters */
+            skipFirstBuild: false,
+            beforeForFirstBuild: false
         };
 
         /** @type {RemovingParameters} */
@@ -207,19 +255,66 @@ class Plugin {
             ...defaultParams,
             ...params.after
         };
+
+        /** @type {{[key: string]: WebpackHook}} */
+        this.webpackHooks = {
+            beforeRun: {
+                name: {
+                    v4: 'beforeRun',
+                    v3: 'before-run'
+                },
+                callsCount: 0,
+                increaseCallsCount: () => {
+                    this.webpackHooks.beforeRun.callsCount++;
+                },
+                getParams: () => this.beforeParams
+            },
+            watchRun: {
+                name: {
+                    v4: 'watchRun',
+                    v3: 'watch-run'
+                },
+                callsCount: 0,
+                increaseCallsCount: () => {
+                    this.webpackHooks.watchRun.callsCount++;
+                },
+                getParams: () => {
+                    if (this.webpackHooks.watchRun.callsCount === 0) {
+                        if (this.watchParams.skipFirstBuild) {
+                            return undefined;
+                        } else if (this.watchParams.beforeForFirstBuild) {
+                            return this.webpackHooks.beforeRun.getParams();
+                        }
+                    }
+
+                    return this.watchParams;
+                }
+            },
+            afterEmit: {
+                name: {
+                    v4: 'afterEmit',
+                    v3: 'after-emit'
+                },
+                callsCount: 0,
+                increaseCallsCount: () => {
+                    this.webpackHooks.afterEmit.callsCount++;
+                },
+                getParams: () => this.afterParams
+            }
+        };
     }
 
     /**
      * "This method is called once by the webpack
      * compiler while installing the plugin.".
      *
-     * @param {Compiler} main
+     * @param {Compiler} compiler
      * "Represents the fully configured webpack environment.".
      *
      * @throws
      * Throws an error if webpack is not able to register the plugin.
      */
-    apply(cmplr) {
+    apply(compiler) {
         const debugName = 'apply: ';
 
         this.loggerDebug.add(
@@ -259,73 +354,107 @@ class Plugin {
             `after parameters – "${JSON.stringify(this.afterParams)}"`
         );
 
+        this.applyHook(
+            compiler,
+            this.webpackHooks.beforeRun
+        );
+        this.applyHook(
+            compiler,
+            this.webpackHooks.watchRun
+        );
+        this.applyHook(
+            compiler,
+            this.webpackHooks.afterEmit
+        );
+    }
+
+    /**
+     * Applies hook to webpack compiler.
+     *
+     * @param {Compiler} compiler
+     * "Represents the fully configured webpack environment.".
+     *
+     * @param {WebpackHook} hook
+     * Webpack hook to apply remove method.
+     *
+     * @throws
+     * Throws an error if webpack is not able to register the plugin.
+     */
+    applyHook(compiler, hook) {
+        const debugName = 'apply-hook: ';
+
         /**
-         * webpack 4+ comes with a new plugin system.
-         * Checking for hooks in order to support old plugin system.
+         * This method will be called by webpack.
+         *
+         * @param {Compiler | Compilation} compilerOrCompilation
+         * Webpack compiler object or compilation object.
+         * It depends on hook.
+         *
+         * @param {Function} callback
+         * "Some compilation plugin steps are asynchronous,
+         * and pass a callback function that must be invoked
+         * when your plugin is finished running".
          */
-        const applyHook = (compiler, v4Hook, v3Hook, params) => {
+        const method = (compilerOrCompilation, callback) => {
+            const params = hook.getParams();
+            const cllbck = () => {
+                hook.increaseCallsCount();
+                callback();
+            };
+
             if (!params || !Object.keys(params).length) {
-                this.loggerDebug.add(
-                    debugName +
-                    'skipped registering, because "params" is empty'
-                );
+                cllbck();
 
                 return;
             }
 
-            const method = (compilerOrCompilation, callback) => {
-                /**
-                 * `params` will be changed during the program lifecycle.
-                 * It is bad pattern, but intended behavior because of
-                 * bad architecture. It is fine for single-run hooks like
-                 * `before` or `after`, but for `watch` hook it can lose
-                 * data in `include` and `exclude`.
-                 */
-                const paramsCopy = {
-                    ...params
-                };
-
-                this.loggerDebug.add(`${debugName}hook started – "${v4Hook}"`);
-                this.handleHook(paramsCopy, callback);
-                this.loggerDebug.add(`${debugName}hook ended – "${v4Hook}"`);
-                this.log(compilerOrCompilation, paramsCopy);
-                this.clear();
+            /**
+             * `params` will be changed during the program lifecycle.
+             * It is bad pattern, but intended behavior because of
+             * bad architecture. It is fine for single-run hooks like
+             * `before` or `after`, but for `watch` hook it can lose
+             * data in `include` and `exclude`.
+             */
+            const paramsCopy = {
+                ...params
             };
 
-            if (compiler.hooks) {
-                compiler.hooks[v4Hook].tapAsync(Info.fullName, method);
-                this.loggerDebug.add(
-                    debugName +
-                    `v4 hook registered – "${v4Hook}"`
-                );
-            } else if (compiler.plugin) {
-                compiler.plugin(v3Hook, method);
-                this.loggerDebug.add(
-                    debugName +
-                    `v3 hook registered – "${v3Hook}"`
-                );
-            } else {
-                throw new Error('webpack is not able to register the plugin');
-            }
+            this.loggerDebug.add(`${debugName}hook started – "${hook.name.v4}"`);
+            this.handleHook(paramsCopy);
+            this.loggerDebug.add(`${debugName}hook ended – "${hook.name.v4}"`);
+            this.log(compilerOrCompilation, paramsCopy);
+            this.clear();
+            cllbck();
         };
 
-        applyHook(cmplr, 'beforeRun', 'before-run', this.beforeParams);
-        applyHook(cmplr, 'watchRun', 'watch-run', this.watchParams);
-        applyHook(cmplr, 'afterEmit', 'after-emit', this.afterParams);
+        /**
+         * webpack 4+ comes with a new plugin system.
+         * Checking for `hooks` in order to support old plugin system.
+         */
+        if (compiler.hooks) {
+            compiler.hooks[hook.name.v4].tapAsync(Info.fullName, method);
+            this.loggerDebug.add(
+                debugName +
+                `v4 hook registered – "${hook.name.v4}"`
+            );
+        } else if (compiler.plugin) {
+            compiler.plugin(hook.name.v3, method);
+            this.loggerDebug.add(
+                debugName +
+                `v3 hook registered – "${hook.name.v3}"`
+            );
+        } else {
+            throw new Error('webpack is not able to register the plugin');
+        }
     }
 
     /**
-     * Handles webpack hooks.
+     * Handles webpack hook.
      *
      * @param {RemovingParameters} params
      * A parameters of removing.
-     *
-     * @param {Function} callback
-     * "Some compilation plugin steps are asynchronous,
-     * and pass a callback function that must be invoked
-     * when your plugin is finished running".
      */
-    handleHook(params, callback) {
+    handleHook(params) {
         const debugName = 'handle-hook: ';
 
         this.path.type = params._pathType;
@@ -335,8 +464,6 @@ class Plugin {
             `path – "${this.path.type || 'auto'}"`
         );
         this.handleRemove(params);
-
-        callback();
     }
 
     /**