+++ types/index.d.ts
@@ -46,6 +46,8 @@ interface RemoveParameters {
      *
      * Defaults to `.` (where package.json and
      * node_modules are located).
+     * 
+     * Namespace: all.
      */
     root?: string;
 
@@ -53,6 +55,8 @@ interface RemoveParameters {
      * A folders and files for removing.
      *
      * Defaults to `[]`.
+     * 
+     * Namespace: all.
      */
     include?: ReadonlyArray<string>;
 
@@ -60,6 +64,8 @@ interface RemoveParameters {
      * A folders and files for excluding.
      *
      * Defaults to `[]`.
+     * 
+     * Namespace: all.
      */
     exclude?: ReadonlyArray<string>;
 
@@ -67,6 +73,8 @@ interface RemoveParameters {
      * A folders for testing.
      *
      * Defaults to `[]`.
+     * 
+     * Namespace: all.
      */
     test?: ReadonlyArray<TestObject>;
 
@@ -76,8 +84,12 @@ interface RemoveParameters {
      * will be passed into this function.
      * If returned value is `true`, then
      * remove process will be canceled.
-     * Will be not called if `emulate` is on.
+     * Will be not called if items for removing 
+     * not found, `emulate: true` or `skipFirstBuild: true`.
+     * 
      * Defaults to `undefined`.
+     * 
+     * Namespace: all.
      */
     beforeRemove?: (
         absoluteFoldersPaths: string[],
@@ -86,9 +98,13 @@ interface RemoveParameters {
 
     /**
      * If specified, will be called after removing.
-     * Absolute paths of folders and files that have been removed
-     * will be passed into this function.
+     * Absolute paths of folders and files that have been 
+     * removed will be passed into this function.
+     * Will be not called if `emulate: true` or `skipFirstBuild: true`.
+     * 
      * Defaults to `undefined`.
+     * 
+     * Namespace: all.
      */
     afterRemove?: (
         absoluteFoldersPaths: string[],
@@ -100,6 +116,8 @@ interface RemoveParameters {
      * instead of permanent removing.
      *
      * Defaults to `false`.
+     * 
+     * Namespace: all.
      */
     trash?: boolean;
 
@@ -108,6 +126,8 @@ interface RemoveParameters {
      * (example: "Which folders or files have been removed").
      *
      * Defaults to `true`.
+     * 
+     * Namespace: all.
      */
     log?: boolean;
 
@@ -116,6 +136,8 @@ interface RemoveParameters {
      * (example: "An items for removing not found").
      *
      * Defaults to `true`.
+     * 
+     * Namespace: all.
      */
     logWarning?: boolean;
 
@@ -124,6 +146,8 @@ interface RemoveParameters {
      * (example: "No such file or directory").
      *
      * Defaults to `false`.
+     * 
+     * Namespace: all.
      */
     logError?: boolean;
 
@@ -132,6 +156,8 @@ interface RemoveParameters {
      * (used for debugging).
      *
      * Defaults to `false`.
+     * 
+     * Namespace: all.
      */
     logDebug?: boolean;
 
@@ -142,6 +168,8 @@ interface RemoveParameters {
      * Ignores `log` parameter.
      *
      * Defaults to `false`.
+     * 
+     * Namespace: all.
      */
     emulate?: boolean;
 
@@ -151,8 +179,29 @@ interface RemoveParameters {
      * **Don't turn it on if you don't know what you actually do!**
      *
      * Defaults to `false`.
+     * 
+     * Namespace: all.
      */
     allowRootAndOutside?: boolean;
+
+    /**
+     * First build will be skipped.
+     * 
+     * Defaults to `false`.
+     * 
+     * Namespace: `watch`.
+     */
+    skipFirstBuild?: boolean;
+
+    /**
+     * For first build `before` parameters will be applied,
+     * for subsequent builds `watch` parameters will be applied.
+     * 
+     * Defaults to `false`.
+     * 
+     * Namespace: `watch`.
+     */
+    beforeForFirstBuild?: boolean;
 }
 
 /**
@@ -161,12 +210,16 @@ interface RemoveParameters {
 interface TestObject {
     /**
      * A path to the folder (relative to `root`).
+     * 
+     * Namespace: all.
      */
     folder: string;
 
     /**
      * A method that accepts an item path (`root` + folderPath + fileName) and
      * returns value that indicates should this item be removed or not.
+     * 
+     * Namespace: all.
      */
     method: (absolutePath: string) => boolean;
 
@@ -174,6 +227,8 @@ interface TestObject {
      * Apply this method to all items in subdirectories.
      *
      * Defaults to `false`.
+     * 
+     * Namespace: all.
      */
     recursive?: boolean;
 }